feat: 출/퇴근 기록 수정/생성 관련 서비스 테스트 생성
